dictionary cleanup
