+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/DictionaryFactory.java
@@ -72,7 +72,6 @@ public interface DictionaryFactory {
 			default:
 				return MatrixBlockDictionary.read(in);
 		}
-
 	}
 
 	public static long getInMemorySize(int nrValues, int nrColumns, double tupleSparsity, boolean lossy) {
