+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/IdentityDictionary.java
@@ -22,7 +22,6 @@ package org.apache.sysds.runtime.compress.colgroup.dictionary;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.lang.ref.SoftReference;
 import java.util.Arrays;
 
 import org.apache.commons.lang3.NotImplementedException;
@@ -36,26 +35,15 @@ import org.apache.sysds.runtime.functionobjects.Builtin.BuiltinCode;
 import org.apache.sysds.runtime.functionobjects.Divide;
 import org.apache.sysds.runtime.functionobjects.Minus;
 import org.apache.sysds.runtime.functionobjects.Plus;
-import org.apache.sysds.runtime.functionobjects.ValueFunction;
-import org.apache.sysds.runtime.instructions.cp.CM_COV_Object;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.runtime.matrix.operators.BinaryOperator;
-import org.apache.sysds.runtime.matrix.operators.ScalarOperator;
-import org.apache.sysds.runtime.matrix.operators.UnaryOperator;
 
 /**
  * A specialized dictionary that exploits the fact that the contained dictionary is an Identity Matrix.
  */
-public class IdentityDictionary extends ADictionary {
+public class IdentityDictionary extends AIdentityDictionary {
 
-	private static final long serialVersionUID = 2535887782150955098L;
-
-	/** The number of rows or columns, rows can be +1 if withEmpty is set. */
-	protected final int nRowCol;
-	/** Specify if the Identity matrix should contain an empty row in the end. */
-	protected final boolean withEmpty;
-	/** A Cache to contain a materialized version of the identity matrix. */
-	protected volatile SoftReference<MatrixBlockDictionary> cache = null;
+	private static final long serialVersionUID = 2535887782153955098L;
 
 	/**
 	 * Create an identity matrix dictionary. It behaves as if allocated a Sparse Matrix block but exploits that the
@@ -64,10 +52,7 @@ public class IdentityDictionary extends ADictionary {
 	 * @param nRowCol The number of rows and columns in this identity matrix.
 	 */
 	public IdentityDictionary(int nRowCol) {
-		if(nRowCol <= 0)
-			throw new DMLCompressionException("Invalid Identity Dictionary");
-		this.nRowCol = nRowCol;
-		this.withEmpty = false;
+		super(nRowCol);
 	}
 
 	/**
@@ -78,10 +63,7 @@ public class IdentityDictionary extends ADictionary {
 	 * @param withEmpty If the matrix should contain an empty row in the end.
 	 */
 	public IdentityDictionary(int nRowCol, boolean withEmpty) {
-		if(nRowCol <= 0)
-			throw new DMLCompressionException("Invalid Identity Dictionary");
-		this.nRowCol = nRowCol;
-		this.withEmpty = withEmpty;
+		super(nRowCol, withEmpty);
 	}
 
 	@Override
@@ -107,10 +89,6 @@ public class IdentityDictionary extends ADictionary {
 		return row == col ? 1 : 0;
 	}
 
-	public boolean withEmpty() {
-		return withEmpty;
-	}
-
 	@Override
 	public double getValue(int r, int c, int nCol) {
 		return r == c ? 1 : 0;
@@ -118,11 +96,11 @@ public class IdentityDictionary extends ADictionary {
 
 	@Override
 	public long getInMemorySize() {
-		return 4 + 4 + 8; // int + padding + softReference
+		return getInMemorySize(-1); // int + padding + softReference
 	}
 
 	public static long getInMemorySize(int numberColumns) {
-		return 4 + 4 + 8;
+		return AIdentityDictionary.getInMemorySize(numberColumns);
 	}
 
 	@Override
@@ -136,25 +114,10 @@ public class IdentityDictionary extends ADictionary {
 	}
 
 	@Override
-	public double aggregateWithReference(double init, Builtin fn, double[] reference, boolean def) {
-		return getMBDict().aggregateWithReference(init, fn, reference, def);
-	}
-
-	@Override
 	public double[] aggregateRows(Builtin fn, int nCol) {
 		double[] ret = new double[nRowCol];
 		Arrays.fill(ret, fn.execute(1, 0));
 		return ret;
-	}
-
-	@Override
-	public double[] aggregateRowsWithDefault(Builtin fn, double[] defaultTuple) {
-		return getMBDict().aggregateRowsWithDefault(fn, defaultTuple);
-	}
-
-	@Override
-	public double[] aggregateRowsWithReference(Builtin fn, double[] reference) {
-		return getMBDict().aggregateRowsWithReference(fn, reference);
 	}
 
 	@Override
@@ -166,59 +129,6 @@ public class IdentityDictionary extends ADictionary {
 		}
 	}
 
-	@Override
-	public void aggregateColsWithReference(double[] c, Builtin fn, IColIndex colIndexes, double[] reference,
-		boolean def) {
-		getMBDict().aggregateColsWithReference(c, fn, colIndexes, reference, def);
-	}
-
-	@Override
-	public IDictionary applyScalarOp(ScalarOperator op) {
-		return getMBDict().applyScalarOp(op);
-	}
-
-	@Override
-	public IDictionary applyScalarOpAndAppend(ScalarOperator op, double v0, int nCol) {
-
-		return getMBDict().applyScalarOpAndAppend(op, v0, nCol);
-	}
-
-	@Override
-	public IDictionary applyUnaryOp(UnaryOperator op) {
-		return getMBDict().applyUnaryOp(op);
-	}
-
-	@Override
-	public IDictionary applyUnaryOpAndAppend(UnaryOperator op, double v0, int nCol) {
-		return getMBDict().applyUnaryOpAndAppend(op, v0, nCol);
-	}
-
-	@Override
-	public IDictionary applyScalarOpWithReference(ScalarOperator op, double[] reference, double[] newReference) {
-		return getMBDict().applyScalarOpWithReference(op, reference, newReference);
-	}
-
-	@Override
-	public IDictionary applyUnaryOpWithReference(UnaryOperator op, double[] reference, double[] newReference) {
-		return getMBDict().applyUnaryOpWithReference(op, reference, newReference);
-	}
-
-	@Override
-	public IDictionary binOpLeft(BinaryOperator op, double[] v, IColIndex colIndexes) {
-		return getMBDict().binOpLeft(op, v, colIndexes);
-	}
-
-	@Override
-	public IDictionary binOpLeftAndAppend(BinaryOperator op, double[] v, IColIndex colIndexes) {
-		return getMBDict().binOpLeftAndAppend(op, v, colIndexes);
-	}
-
-	@Override
-	public IDictionary binOpLeftWithReference(BinaryOperator op, double[] v, IColIndex colIndexes, double[] reference,
-		double[] newReference) {
-		return getMBDict().binOpLeftWithReference(op, v, colIndexes, reference, newReference);
-
-	}
 
 	@Override
 	public IDictionary binOpRight(BinaryOperator op, double[] v, IColIndex colIndexes) {
@@ -237,22 +147,6 @@ public class IdentityDictionary extends ADictionary {
 			return this;
 		MatrixBlockDictionary mb = getMBDict();
 		return mb.binOpRight(op, v, colIndexes);
-	}
-
-	@Override
-	public IDictionary binOpRightAndAppend(BinaryOperator op, double[] v, IColIndex colIndexes) {
-		return getMBDict().binOpRightAndAppend(op, v, colIndexes);
-	}
-
-	@Override
-	public IDictionary binOpRight(BinaryOperator op, double[] v) {
-		return getMBDict().binOpRight(op, v);
-	}
-
-	@Override
-	public IDictionary binOpRightWithReference(BinaryOperator op, double[] v, IColIndex colIndexes, double[] reference,
-		double[] newReference) {
-		return getMBDict().binOpRightWithReference(op, v, colIndexes, reference, newReference);
 	}
 
 	@Override
@@ -322,31 +216,6 @@ public class IdentityDictionary extends ADictionary {
 	}
 
 	@Override
-	public double[] sumAllRowsToDoubleSqWithDefault(double[] defaultTuple) {
-		return getMBDict().sumAllRowsToDoubleSqWithDefault(defaultTuple);
-	}
-
-	@Override
-	public double[] sumAllRowsToDoubleSqWithReference(double[] reference) {
-		return getMBDict().sumAllRowsToDoubleSqWithReference(reference);
-	}
-
-	@Override
-	public double[] productAllRowsToDouble(int nCol) {
-		return new double[nRowCol];
-	}
-
-	@Override
-	public double[] productAllRowsToDoubleWithDefault(double[] defaultTuple) {
-		return new double[nRowCol];
-	}
-
-	@Override
-	public double[] productAllRowsToDoubleWithReference(double[] reference) {
-		return getMBDict().productAllRowsToDoubleWithReference(reference);
-	}
-
-	@Override
 	public void colSum(double[] c, int[] counts, IColIndex colIndexes) {
 		for(int i = 0; i < colIndexes.size(); i++)
 			c[colIndexes.get(i)] += counts[i];
@@ -364,16 +233,6 @@ public class IdentityDictionary extends ADictionary {
 		}
 	}
 
-	@Override
-	public void colProductWithReference(double[] res, int[] counts, IColIndex colIndexes, double[] reference) {
-		getMBDict().colProductWithReference(res, counts, colIndexes, reference);
-
-	}
-
-	@Override
-	public void colSumSqWithReference(double[] c, int[] counts, IColIndex colIndexes, double[] reference) {
-		getMBDict().colSumSqWithReference(c, counts, colIndexes, reference);
-	}
 
 	@Override
 	public double sum(int[] counts, int ncol) {
@@ -392,26 +251,11 @@ public class IdentityDictionary extends ADictionary {
 	}
 
 	@Override
-	public double sumSqWithReference(int[] counts, double[] reference) {
-		return getMBDict().sumSqWithReference(counts, reference);
-	}
-
-	@Override
 	public IDictionary sliceOutColumnRange(int idxStart, int idxEnd, int previousNumberOfColumns) {
 		if(idxStart == 0 && idxEnd == nRowCol)
 			return new IdentityDictionary(nRowCol, withEmpty);
 		else
 			return new IdentityDictionarySlice(nRowCol, withEmpty, idxStart, idxEnd);
-	}
-
-	@Override
-	public boolean containsValue(double pattern) {
-		return pattern == 0.0 || pattern == 1.0;
-	}
-
-	@Override
-	public boolean containsValueWithReference(double pattern, double[] reference) {
-		return getMBDict().containsValueWithReference(pattern, reference);
 	}
 
 	@Override
@@ -422,11 +266,6 @@ public class IdentityDictionary extends ADictionary {
 	@Override
 	public int[] countNNZZeroColumns(int[] counts) {
 		return counts; // interesting ... but true.
-	}
-
-	@Override
-	public long getNumberNonZerosWithReference(int[] counts, double[] reference, int nRows) {
-		return getMBDict().getNumberNonZerosWithReference(counts, reference, nRows);
 	}
 
 	@Override
@@ -483,28 +322,13 @@ public class IdentityDictionary extends ADictionary {
 		v[t8 * nCol + f8] += 1;
 	}
 
-	@Override
-	public IDictionary subtractTuple(double[] tuple) {
-		return getMBDict().subtractTuple(tuple);
-	}
-
-	public MatrixBlockDictionary getMBDict() {
+	@Override 
+	public MatrixBlockDictionary getMBDict(){
 		return getMBDict(nRowCol);
 	}
 
 	@Override
-	public MatrixBlockDictionary getMBDict(int nCol) {
-		if(cache != null) {
-			MatrixBlockDictionary r = cache.get();
-			if(r != null)
-				return r;
-		}
-		MatrixBlockDictionary ret = createMBDict();
-		cache = new SoftReference<>(ret);
-		return ret;
-	}
-
-	private MatrixBlockDictionary createMBDict() {
+	public MatrixBlockDictionary createMBDict(int nCol) {
 
 		if(withEmpty) {
 			final SparseBlock sb = SparseBlockFactory.createIdentityMatrixWithEmptyRow(nRowCol);
@@ -519,10 +343,6 @@ public class IdentityDictionary extends ADictionary {
 		}
 	}
 
-	@Override
-	public IDictionary scaleTuples(int[] scaling, int nCol) {
-		return getMBDict().scaleTuples(scaling, nCol);
-	}
 
 	@Override
 	public void write(DataOutput out) throws IOException {
@@ -567,64 +387,6 @@ public class IdentityDictionary extends ADictionary {
 	}
 
 	@Override
-	public IDictionary replace(double pattern, double replace, int nCol) {
-		if(containsValue(pattern))
-			return getMBDict().replace(pattern, replace, nCol);
-		else
-			return this;
-	}
-
-	@Override
-	public IDictionary replaceWithReference(double pattern, double replace, double[] reference) {
-		if(containsValueWithReference(pattern, reference))
-			return getMBDict().replaceWithReference(pattern, replace, reference);
-		else
-			return this;
-	}
-
-	@Override
-	public void product(double[] ret, int[] counts, int nCol) {
-		getMBDict().product(ret, counts, nCol);
-	}
-
-	@Override
-	public void productWithDefault(double[] ret, int[] counts, double[] def, int defCount) {
-		getMBDict().productWithDefault(ret, counts, def, defCount);
-	}
-
-	@Override
-	public void productWithReference(double[] ret, int[] counts, double[] reference, int refCount) {
-		getMBDict().productWithReference(ret, counts, reference, refCount);
-	}
-
-	@Override
-	public CM_COV_Object centralMoment(CM_COV_Object ret, ValueFunction fn, int[] counts, int nRows) {
-		return getMBDict().centralMoment(ret, fn, counts, nRows);
-	}
-
-	@Override
-	public CM_COV_Object centralMomentWithDefault(CM_COV_Object ret, ValueFunction fn, int[] counts, double def,
-		int nRows) {
-		return getMBDict().centralMomentWithDefault(ret, fn, counts, def, nRows);
-	}
-
-	@Override
-	public CM_COV_Object centralMomentWithReference(CM_COV_Object ret, ValueFunction fn, int[] counts, double reference,
-		int nRows) {
-		return getMBDict().centralMomentWithReference(ret, fn, counts, reference, nRows);
-	}
-
-	@Override
-	public IDictionary rexpandCols(int max, boolean ignore, boolean cast, int nCol) {
-		return getMBDict().rexpandCols(max, ignore, cast, nCol);
-	}
-
-	@Override
-	public IDictionary rexpandColsWithReference(int max, boolean ignore, boolean cast, int reference) {
-		return getMBDict().rexpandColsWithReference(max, ignore, cast, reference);
-	}
-
-	@Override
 	public double getSparsity() {
 		if(withEmpty)
 			return 1d / (nRowCol + 1);
@@ -636,21 +398,6 @@ public class IdentityDictionary extends ADictionary {
 	public void multiplyScalar(double v, double[] ret, int off, int dictIdx, IColIndex cols) {
 		if(!withEmpty || dictIdx < nRowCol)
 			ret[off + cols.get(dictIdx)] += v;
-	}
-
-	@Override
-	public void TSMMWithScaling(int[] counts, IColIndex rows, IColIndex cols, MatrixBlock ret) {
-		getMBDict().TSMMWithScaling(counts, rows, cols, ret);
-	}
-
-	@Override
-	public void MMDict(IDictionary right, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result) {
-		getMBDict().MMDict(right, rowsLeft, colsRight, result);
-	}
-
-	public void MMDictScaling(IDictionary right, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result,
-		int[] scaling) {
-		getMBDict().MMDictScaling(right, rowsLeft, colsRight, result, scaling);
 	}
 
 	@Override
@@ -673,7 +420,6 @@ public class IdentityDictionary extends ADictionary {
 	@Override
 	public void MMDictScalingDense(double[] left, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result,
 		int[] scaling) {
-		// getMBDict().MMDictScalingDense(left, rowsLeft, colsRight, result, scaling);
 		final int leftSide = rowsLeft.size();
 		final int resCols = result.getNumColumns();
 		final double[] resV = result.getDenseBlockValues();
@@ -685,51 +431,6 @@ public class IdentityDictionary extends ADictionary {
 		}
 	}
 
-	@Override
-	public void MMDictSparse(SparseBlock left, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result) {
-		getMBDict().MMDictSparse(left, rowsLeft, colsRight, result);
-	}
-
-	@Override
-	public void MMDictScalingSparse(SparseBlock left, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result,
-		int[] scaling) {
-		getMBDict().MMDictScalingSparse(left, rowsLeft, colsRight, result, scaling);
-	}
-
-	@Override
-	public void TSMMToUpperTriangle(IDictionary right, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result) {
-		getMBDict().TSMMToUpperTriangle(right, rowsLeft, colsRight, result);
-	}
-
-	@Override
-	public void TSMMToUpperTriangleDense(double[] left, IColIndex rowsLeft, IColIndex colsRight, MatrixBlock result) {
-		getMBDict().TSMMToUpperTriangleDense(left, rowsLeft, colsRight, result);
-	}
-
-	@Override
-	public void TSMMToUpperTriangleSparse(SparseBlock left, IColIndex rowsLeft, IColIndex colsRight,
-		MatrixBlock result) {
-		getMBDict().TSMMToUpperTriangleSparse(left, rowsLeft, colsRight, result);
-	}
-
-	@Override
-	public void TSMMToUpperTriangleScaling(IDictionary right, IColIndex rowsLeft, IColIndex colsRight, int[] scale,
-		MatrixBlock result) {
-		getMBDict().TSMMToUpperTriangleScaling(right, rowsLeft, colsRight, scale, result);
-	}
-
-	@Override
-	public void TSMMToUpperTriangleDenseScaling(double[] left, IColIndex rowsLeft, IColIndex colsRight, int[] scale,
-		MatrixBlock result) {
-		getMBDict().TSMMToUpperTriangleDenseScaling(left, rowsLeft, colsRight, scale, result);
-	}
-
-	@Override
-	public void TSMMToUpperTriangleSparseScaling(SparseBlock left, IColIndex rowsLeft, IColIndex colsRight, int[] scale,
-		MatrixBlock result) {
-
-		getMBDict().TSMMToUpperTriangleSparseScaling(left, rowsLeft, colsRight, scale, result);
-	}
 
 	@Override
 	public boolean equals(IDictionary o) {
@@ -738,16 +439,6 @@ public class IdentityDictionary extends ADictionary {
 			((IdentityDictionary) o).withEmpty == withEmpty)
 			return true;
 		return getMBDict().equals(o);
-	}
-
-	@Override
-	public IDictionary cbind(IDictionary that, int nCol) {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public IDictionary reorder(int[] reorder) {
-		return getMBDict().reorder(reorder);
 	}
 
 	@Override
@@ -810,11 +501,6 @@ public class IdentityDictionary extends ADictionary {
 		final MatrixBlock retB = new MatrixBlock(numVals, aggregateColumns.size(), -1, ret);
 		retB.recomputeNonZeros();
 		return MatrixBlockDictionary.create(retB, false);
-	}
-
-	@Override
-	public IDictionary append(double[] row) {
-		return getMBDict().append(row);
 	}
 
 	@Override
