+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/IdentityDictionarySlice.java
@@ -22,20 +22,16 @@ package org.apache.sysds.runtime.compress.colgroup.dictionary;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.lang.ref.SoftReference;
 import java.util.Arrays;
 
-import org.apache.commons.lang3.NotImplementedException;
 import org.apache.sysds.runtime.DMLRuntimeException;
 import org.apache.sysds.runtime.compress.colgroup.indexes.IColIndex;
-import org.apache.sysds.runtime.data.SparseBlock;
 import org.apache.sysds.runtime.functionobjects.Builtin;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
-import org.apache.sysds.runtime.matrix.operators.BinaryOperator;
 
-public class IdentityDictionarySlice extends IdentityDictionary {
+public class IdentityDictionarySlice extends AIdentityDictionary {
 
-	private static final long serialVersionUID = 2535887782150955098L;
+	private static final long serialVersionUID = 2535887782153555098L;
 
 	/** Lower index for the slice */
 	private final int l;
@@ -72,14 +68,20 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 
 	@Override
 	public double getValue(int i) {
-		throw new NotImplementedException();
+		final int nCol = u - l;
+		final int vRow = i / nCol;
+		if(vRow < l || vRow >= u)
+			return 0;
+		final int oRow = vRow - l;
+		final int col = i % nCol;
+		return oRow == col ? 1 : 0;
 	}
 
 	@Override
 	public final double getValue(int r, int c, int nCol) {
 		if(r < l || r > u)
 			return 0;
-		return super.getValue(r - l, c, nCol);
+		return (r - l) == c ? 1 : 0;
 	}
 
 	@Override
@@ -88,8 +90,8 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 	}
 
 	public static long getInMemorySize(int numberColumns) {
-		// int * 3 + padding + softReference
-		return 12 + 4 + 8;
+		// 2 more ints, no padding.
+		return AIdentityDictionary.getInMemorySize(numberColumns) + 8;
 	}
 
 	@Override
@@ -127,19 +129,29 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 
 	@Override
 	public double[] sumAllRowsToDoubleWithDefault(double[] defaultTuple) {
-		double[] ret = new double[nRowCol];
-		Arrays.fill(ret, l, u, 1.0);
+		double[] ret = new double[getNumberOfValues(defaultTuple.length) + 1];
+		for(int i = l; i < u; i++)
+			ret[i] = 1;
 		for(int i = 0; i < defaultTuple.length; i++)
-			ret[i] += defaultTuple[i];
+			ret[ret.length - 1] += defaultTuple[i];
 		return ret;
 	}
 
 	@Override
 	public double[] sumAllRowsToDoubleWithReference(double[] reference) {
-		double[] ret = new double[nRowCol];
-		Arrays.fill(ret, l, u, 1.0);
+		double[] ret = new double[getNumberOfValues(reference.length)];
+		double refSum = 0;
 		for(int i = 0; i < reference.length; i++)
-			ret[i] += reference[i] * nRowCol;
+			refSum += reference[i];
+		for(int i = 0; i < ret.length; i++) {
+			if(i < l || i > u)
+				ret[i] = refSum;
+			else
+				ret[i] = 1 + refSum;
+		}
+
+		if(withEmpty)
+			ret[ret.length - 1] += -1;
 		return ret;
 	}
 
@@ -162,11 +174,8 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 
 	@Override
 	public void colSum(double[] c, int[] counts, IColIndex colIndexes) {
-		for(int i = 0; i < colIndexes.size(); i++) {
-			// very nice...
-			final int idx = colIndexes.get(i);
-			c[idx] = counts[i];
-		}
+		for(int i = l; i < u; i++)
+			c[colIndexes.get(i - l)] = counts[i];
 	}
 
 	@Override
@@ -184,16 +193,6 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 	}
 
 	@Override
-	public IDictionary sliceOutColumnRange(int idxStart, int idxEnd, int previousNumberOfColumns) {
-		return getMBDict().sliceOutColumnRange(idxStart, idxEnd, previousNumberOfColumns);
-	}
-
-	@Override
-	public boolean containsValue(double pattern) {
-		return pattern == 0.0 || pattern == 1.0;
-	}
-
-	@Override
 	public long getNumberNonZeros(int[] counts, int nCol) {
 		return (long) sum(counts, nCol);
 	}
@@ -201,40 +200,6 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 	@Override
 	public int getNumberOfValues(int ncol) {
 		return nRowCol + (withEmpty ? 1 : 0);
-	}
-
-	@Override
-	public MatrixBlockDictionary getMBDict(int nCol) {
-		if(cache != null) {
-			MatrixBlockDictionary r = cache.get();
-			if(r != null)
-				return r;
-		}
-		MatrixBlockDictionary ret = createMBDict();
-		cache = new SoftReference<>(ret);
-		return ret;
-	}
-
-	private MatrixBlockDictionary createMBDict() {
-		MatrixBlock identity = new MatrixBlock(nRowCol, u - l, true);
-		for(int i = l; i < u; i++)
-			identity.set(i, i - l, 1.0);
-		return new MatrixBlockDictionary(identity);
-	}
-
-	@Override
-	public String getString(int colIndexes) {
-		return "IdentityMatrix of size: " + nRowCol;
-	}
-
-	@Override
-	public String toString() {
-		return "IdentityMatrix of size: " + nRowCol;
-	}
-
-	@Override
-	public IDictionary scaleTuples(int[] scaling, int nCol) {
-		return getMBDict().scaleTuples(scaling, nCol);
 	}
 
 	@Override
@@ -262,46 +227,13 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 	}
 
 	@Override
-	public IDictionary replace(double pattern, double replace, int nCol) {
-		if(containsValue(pattern))
-			return getMBDict().replace(pattern, replace, nCol);
-		else
-			return this;
-	}
-
-	@Override
-	public IDictionary replaceWithReference(double pattern, double replace, double[] reference) {
-		if(containsValueWithReference(pattern, reference))
-			return getMBDict().replaceWithReference(pattern, replace, reference);
-		else
-			return this;
-	}
-
-	@Override
 	public double getSparsity() {
-		return (double) (u - l) / ((u -l) * (nRowCol + (withEmpty ? 1 : 0)));
-	}
-
-	@Override
-	public IDictionary binOpRight(BinaryOperator op, double[] v, IColIndex colIndexes) {
-		return getMBDict().binOpRight(op, v);
-	}
-
-	@Override
-	public IDictionary preaggValuesFromDense(final int numVals, final IColIndex colIndexes,
-		final IColIndex aggregateColumns, final double[] b, final int cut) {
-		return getMBDict().preaggValuesFromDense(numVals, colIndexes, aggregateColumns, b, cut);
-	}
-
-	@Override
-	public void addToEntryVectorized(double[] v, int f1, int f2, int f3, int f4, int f5, int f6, int f7, int f8, int t1,
-		int t2, int t3, int t4, int t5, int t6, int t7, int t8, int nCol) {
-		getMBDict().addToEntryVectorized(v, f1, f2, f3, f4, f5, f6, f7, f8, t1, t2, t3, t4, t5, t6, t7, t8, nCol);
+		return (double) (u - l) / ((u - l) * (nRowCol + (withEmpty ? 1 : 0)));
 	}
 
 	@Override
 	public void addToEntry(final double[] v, final int fr, final int to, final int nCol, int rep) {
-		if(fr >= l && fr < u) 
+		if(fr >= l && fr < u)
 			v[to * nCol + fr - l] += rep;
 	}
 
@@ -317,24 +249,27 @@ public class IdentityDictionarySlice extends IdentityDictionary {
 			return getMBDict().equals(o);
 	}
 
-	@Override
-	public MatrixBlockDictionary getMBDict() {
-		final int nCol = u - l;
-		MatrixBlock mb = new MatrixBlock(nRowCol + (withEmpty ? 1 : 0), nCol, true);
-		mb.allocateSparseRowsBlock();
-
-		SparseBlock sb = mb.getSparseBlock();
-		for(int i = l; i < u; i++) {
-			sb.append(i, i - l, 1);
-		}
-
-		mb.setNonZeros(nCol);
-		return new MatrixBlockDictionary(mb);
+	@Override 
+	public MatrixBlockDictionary getMBDict(){
+		return getMBDict(nRowCol);
 	}
 
 	@Override
-	public void multiplyScalar(double v, double[] ret, int off, int dictIdx, IColIndex cols) {
-		getMBDict().multiplyScalar(v, ret, off, dictIdx, cols);
+	public MatrixBlockDictionary createMBDict(int nCol) {
+		MatrixBlock identity = new MatrixBlock(nRowCol + (withEmpty ?  1 : 0), u - l, true);
+		for(int i = l; i < u; i++)
+			identity.set(i, i - l, 1.0);
+		return new MatrixBlockDictionary(identity);
+	}
+
+	@Override
+	public String getString(int colIndexes) {
+		return toString();
+	}
+
+	@Override
+	public String toString() {
+		return "IdentityMatrixSlice of size: " + nRowCol + " l " + l + " u " + u;
 	}
 
 }
