+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/MatrixBlockDictionary.java
@@ -2505,8 +2505,8 @@ public class MatrixBlockDictionary extends ADictionary {
 		if(o instanceof MatrixBlockDictionary)
 			return _data.equals(((MatrixBlockDictionary) o)._data);
 
-		else if(o instanceof IdentityDictionary)
-			return ((IdentityDictionary) o).equals(this);
+		else if(o instanceof AIdentityDictionary)
+			return ((AIdentityDictionary) o).equals(this);
 		else if(o instanceof Dictionary) {
 			double[] dVals = ((Dictionary) o)._values;
 			if(_data.isEmpty()) {
@@ -2521,7 +2521,7 @@ public class MatrixBlockDictionary extends ADictionary {
 			final double[] dv = _data.getDenseBlockValues();
 			return Arrays.equals(dv, dVals);
 		}
-		else if(o instanceof IdentityDictionary) {
+		else if(o instanceof AIdentityDictionary) {
 			return o.equals(this);
 		}
 
@@ -2588,4 +2588,85 @@ public class MatrixBlockDictionary extends ADictionary {
 			return new MatrixBlockDictionary(mb);
 		}
 	}
+
+	@Override
+	protected IDictionary rightMMPreAggSparseSelectedCols(int numVals, SparseBlock b, IColIndex thisCols,
+		IColIndex aggregateColumns) {
+
+		final int thisColsSize = thisCols.size();
+		final int aggColSize = aggregateColumns.size();
+		final double[] ret = new double[numVals * aggColSize];
+
+		for(int h = 0; h < thisColsSize; h++) {
+			// chose row in right side matrix via column index of the dictionary
+			final int colIdx = thisCols.get(h);
+			if(b.isEmpty(colIdx))
+				continue;
+
+			// extract the row values on the right side.
+			final double[] sValues = b.values(colIdx);
+			final int[] sIndexes = b.indexes(colIdx);
+			final int sPos = b.pos(colIdx);
+			final int sEnd = b.size(colIdx) + sPos;
+
+			for(int j = 0; j < numVals; j++) { // rows left
+				final int offOut = j * aggColSize;
+				final double v = getValue(j, h, thisColsSize);
+				sparseAddSelected(sPos, sEnd, aggColSize, aggregateColumns, sIndexes, sValues, ret, offOut, v);
+			}
+
+		}
+		return Dictionary.create(ret);
+	}
+
+	private void sparseAddSelected(int sPos, int sEnd, int aggColSize, IColIndex aggregateColumns, int[] sIndexes,
+		double[] sValues, double[] ret, int offOut, double v) {
+
+		int retIdx = 0;
+		for(int i = sPos; i < sEnd; i++) {
+			// skip through the retIdx.
+			while(retIdx < aggColSize && aggregateColumns.get(retIdx) < sIndexes[i])
+				retIdx++;
+			if(retIdx == aggColSize)
+				break;
+			ret[offOut + retIdx] += v * sValues[i];
+		}
+		retIdx = 0;
+	}
+
+	@Override
+	protected IDictionary rightMMPreAggSparseAllColsRight(int numVals, SparseBlock b, IColIndex thisCols,
+		int nColRight) {
+		final int thisColsSize = thisCols.size();
+		final double[] ret = new double[numVals * nColRight];
+
+		for(int h = 0; h < thisColsSize; h++) { // common dim
+			// chose row in right side matrix via column index of the dictionary
+			final int colIdx = thisCols.get(h);
+			if(b.isEmpty(colIdx))
+				continue;
+
+			// extract the row values on the right side.
+			final double[] sValues = b.values(colIdx);
+			final int[] sIndexes = b.indexes(colIdx);
+			final int sPos = b.pos(colIdx);
+			final int sEnd = b.size(colIdx) + sPos;
+
+			for(int i = 0; i < numVals; i++) { // rows left
+				final int offOut = i * nColRight;
+				final double v = getValue(i, h, thisColsSize);
+				SparseAdd(sPos, sEnd, ret, offOut, sIndexes, sValues, v);
+			}
+		}
+		return Dictionary.create(ret);
+	}
+
+	private void SparseAdd(int sPos, int sEnd, double[] ret, int offOut, int[] sIdx, double[] sVals, double v) {
+		if(v != 0) {
+			for(int k = sPos; k < sEnd; k++) { // cols right with value
+				ret[offOut + sIdx[k]] += v * sVals[k];
+			}
+		}
+	}
+
 }
